Combined Evghenii's code with my SQLite Tester.
* Code from SQLite Tester now follows Evghenii's lead with containing
  exceptions within the functions.

* Schema in the SQLite Tester functions has been altered to fit Evghenii's
  changes in the DB schema.

* Rigorous testing must be done to ensure that all code actually works. I'll be
  doing a whole bunch tomorrow.

+++ src/Service.java
@@ -1,12 +1,12 @@
 class Service {
     String name;
     int code;
-    Double fee;
+    float fee;
     Boolean status;
 
     // ---- Constructors ----
     // makes service for DB Wrapper
-    public Service(int code, String name, Double fee, int status)
+    public Service(int code, String name, float fee, int status)
         throws InputException {
         String exceptionString = "";
 
@@ -39,7 +39,7 @@ class Service {
     }
 
     // makes new service -- uses first constructor
-    public Service (String name, Double fee)
+    public Service (String name, float fee)
         throws InputException {
         this(0, name, fee, 1);
         return;
@@ -54,7 +54,7 @@ class Service {
         this.code = code;
     }
 
-    private void setFee(Double fee) {
+    private void setFee(float fee) {
         this.fee = fee;
     }
 
@@ -70,7 +70,7 @@ class Service {
         return code;
     }
 
-    public double getFee() {
+    public float getFee() {
         return fee;
     }
